Moves hooks
