+++ packages/server/src/models/index.js
@@ -1,6 +1,6 @@
 const { getDB } = require('./db');
 const { init } = require('./init');
-const { addHooks } = require('./hooks');
+const { addHooks } = require('../sync/hooks');
 
 const db = getDB();
 init(db);
